Solution structure rework, gitattributes, and gitignore
