+++ .gitattributes
@@ -1,0 +1,2 @@
+# Noralize line endings
+* text=auto
