+++ .gitignore
@@ -1,0 +1,3 @@
+# Ignore output and intermediate build output
+bin/
+obj/
