--- TeaExample/Example.cs
@@ -1,8 +1,0 @@
-﻿using TeaFramework;
-
-namespace TeaExample
-{
-    public class Example : TeaMod
-    {
-    }
-}
