--- TeaFramework.csproj
@@ -1,32 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<Project Sdk="Microsoft.NET.Sdk">
-  <Import Project="..\tModLoader.targets" />
-  <PropertyGroup>
-    <AssemblyName>TeaFramework</AssemblyName>
-    <TargetFramework>net6.0</TargetFramework>
-    <PlatformTarget>AnyCPU</PlatformTarget>
-    <LangVersion>latest</LangVersion>
-    <Nullable>enable</Nullable>
-  </PropertyGroup>
-  <ItemGroup>
-    <PackageReference Include="JetBrains.Annotations" Version="2021.3.0" />
-    <PackageReference Include="tModLoader.CodeAssist" Version="0.1.*" />
-  </ItemGroup>
-  <ItemGroup>
-    <Reference Include="Tomlet, Version=1.3.5.0, Culture=neutral, PublicKeyToken=null">
-      <HintPath>lib\Tomlet.dll</HintPath>
-    </Reference>
-  </ItemGroup>
-  <ItemGroup>
-    <Compile Remove="TeaFramework.Tests\**" />
-    <Compile Remove="TeaExample\**" />
-  </ItemGroup>
-  <ItemGroup>
-    <EmbeddedResource Remove="TeaFramework.Tests\**" />
-    <EmbeddedResource Remove="TeaExample\**" />
-  </ItemGroup>
-  <ItemGroup>
-    <None Remove="TeaFramework.Tests\**" />
-    <None Remove="TeaExample\**" />
-  </ItemGroup>
-</Project>
