--- TeaFramework.sln
@@ -1,34 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.30114.105
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TeaFramework", "TeaFramework.csproj", "{EFF24DB2-170C-47E2-BC8C-4C40FB4D3D9D}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TeaFramework.Tests", "TeaFramework.Tests\TeaFramework.Tests.csproj", "{323CB1AE-E78C-4089-93B4-3B066494370E}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TeaExample", "TeaExample\TeaExample.csproj", "{BC0EBEFD-26E4-414D-8D86-6D65C00DFAB4}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{EFF24DB2-170C-47E2-BC8C-4C40FB4D3D9D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{EFF24DB2-170C-47E2-BC8C-4C40FB4D3D9D}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{EFF24DB2-170C-47E2-BC8C-4C40FB4D3D9D}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{EFF24DB2-170C-47E2-BC8C-4C40FB4D3D9D}.Release|Any CPU.Build.0 = Release|Any CPU
-		{323CB1AE-E78C-4089-93B4-3B066494370E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{323CB1AE-E78C-4089-93B4-3B066494370E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{323CB1AE-E78C-4089-93B4-3B066494370E}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{323CB1AE-E78C-4089-93B4-3B066494370E}.Release|Any CPU.Build.0 = Release|Any CPU
-		{BC0EBEFD-26E4-414D-8D86-6D65C00DFAB4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{BC0EBEFD-26E4-414D-8D86-6D65C00DFAB4}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{BC0EBEFD-26E4-414D-8D86-6D65C00DFAB4}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{BC0EBEFD-26E4-414D-8D86-6D65C00DFAB4}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-EndGlobal
