--- Utility/Cecil.cs
@@ -1,9 +1,0 @@
-﻿namespace TeaFramework.Utility
-{
-    /// <summary>
-    ///     Various <c>Mono.Cecil</c>-related utilities.
-    /// </summary>
-    public static partial class Cecil
-    {
-    }
-}
