+++ src/TeaExample/Example.cs
@@ -1,0 +1,8 @@
+﻿using TeaFramework;
+
+namespace TeaExampleMod
+{
+    public class Example : TeaMod
+    {
+    }
+}
