+++ src/TeaFramework.sln
@@ -1,0 +1,42 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.1.32414.318
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TeaFramework", "TeaFramework\TeaFramework.csproj", "{EFF24DB2-170C-47E2-BC8C-4C40FB4D3D9D}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TeaFramework.Tests", "TeaFramework.Tests\TeaFramework.Tests.csproj", "{323CB1AE-E78C-4089-93B4-3B066494370E}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TeaExampleMod", "TeaExample\TeaExampleMod.csproj", "{BC0EBEFD-26E4-414D-8D86-6D65C00DFAB4}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{365511C4-F690-4878-B2FA-6183F9CA099B}"
+	ProjectSection(SolutionItems) = preProject
+		.editorconfig = .editorconfig
+	EndProjectSection
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{EFF24DB2-170C-47E2-BC8C-4C40FB4D3D9D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{EFF24DB2-170C-47E2-BC8C-4C40FB4D3D9D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{EFF24DB2-170C-47E2-BC8C-4C40FB4D3D9D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{EFF24DB2-170C-47E2-BC8C-4C40FB4D3D9D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{323CB1AE-E78C-4089-93B4-3B066494370E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{323CB1AE-E78C-4089-93B4-3B066494370E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{323CB1AE-E78C-4089-93B4-3B066494370E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{323CB1AE-E78C-4089-93B4-3B066494370E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{BC0EBEFD-26E4-414D-8D86-6D65C00DFAB4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{BC0EBEFD-26E4-414D-8D86-6D65C00DFAB4}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{BC0EBEFD-26E4-414D-8D86-6D65C00DFAB4}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{BC0EBEFD-26E4-414D-8D86-6D65C00DFAB4}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {F7E245F3-9249-4969-B2DF-6FE5D097A90A}
+	EndGlobalSection
+EndGlobal
