+++ src/TeaFramework/API/Patching/DetourPatch.cs
