+++ src/TeaFramework/API/Patching/IMonoModPatch.cs
