+++ src/TeaFramework/Impl/Utility/Cecil.cs
@@ -1,0 +1,9 @@
+﻿namespace TeaFramework.Impl.Utility
+{
+    /// <summary>
+    ///     Various <c>Mono.Cecil</c>-related utilities.
+    /// </summary>
+    public static partial class Cecil
+    {
+    }
+}
